Parameter Make.C

val read_frontend_mac : S.id -> Macaddr.t Lwt.t

val read_backend_mac : S.id -> Macaddr.t Lwt.t

val read_mtu : S.id -> int Lwt.t

val read_frontend_configuration : S.id -> S.frontend_configuration Lwt.t

Waits for the frontend configuration to become available and returns it.

val write_frontend_configuration : S.id -> S.frontend_configuration -> unit Lwt.t

val enumerate : unit -> string list Lwt.t

List the names of available devices.

val connect : S.id -> unit Lwt.t

val disconnect_frontend : S.id -> unit Lwt.t

Set the frontend state to Closed.

val disconnect_backend : S.id -> unit Lwt.t

Delete the backend directory.

val wait_until_backend_connected : S.backend_configuration -> unit Lwt.t

val read_backend : S.id -> S.backend_configuration Lwt.t

val init_backend : S.id -> Features.t -> S.backend_configuration Lwt.t

Initialise the configuration for a new backend.

val description : string

Human-readable description suitable for help text or a manpage

val wait_for_frontend_closing : S.id -> unit Lwt.t

wait_for_frontend_closing id is a thread that returns when id's frontend moves to the closing state.

val wait_for_backend_closing : S.id -> unit Lwt.t

wait_for_backend_closing id is a thread that returns when id's backend moves to the closing state.