Module Stats

val create : unit -> Mirage_net.stats

create () returns a fresh set of zeroed counters

val rx : Mirage_net.stats -> int64 -> unit

rx t size records that we received a packet of length size

val tx : Mirage_net.stats -> int64 -> unit

tx t size records that we transmitted a packet of length size

val reset : Mirage_net.stats -> unit

reset t resets all packet counters in t to 0