mirage-net-xen index

Library mirage-net-xen

This library exposes the following toplevel modules:

Assemble Assemble complete network frames from Xen network messages.

Backend

Features

Flags

Frontend

Netif

RX

S

Shared_page_pool Pool and reuse shared memory blocks. This saves having to reallocate and reshare pages of memory, which is slow. A block is half the size of a page, since network packets usually don't require more than this.

Stats

TX

Xenstore